refactor: 재사용성이 낮은 useParticipationButton을 삭제하고, 해당 로직을 DetailCard 컴포넌트 내부에서 관리하도록 수정
